// src/app/Models/NewReturns.model.ts
export class NewReturns {
    MailType: string;
    CampaignName: string;
    PhaseName: string;
    MailCode: string;
    MailDescription: string;
    Mailed: Date;
    Caged: Date;
    Quantity: number;
    Donors: number;
    NonDonors: number;
    NewDonors: number;
    RSP: number;
    AVG: number;
    Gross: number;
    Cost: number;
    Net: number;
    GPP: number;
    CLM: number;
    NLM: number;
    IO: number;
}